Refactors IScene interface
Moves the IScene interface definition from a dedicated file to the RenderableObjects module, replacing the previous IGeometry definition.
This change reorganizes the project structure, grouping scene-related functionalities within the RenderableObjects module for better cohesion.
Additionally, the destructor of the pure interfaces are now defined out-of-line to avoid compilation issues when using them as members in other classes.