Moves Vertex definition to PhysicalProperties
Moves the Vertex definition from the Rendering module to the PhysicalProperties module.

This change aligns with the architectural shift to make primitive
core data structures part of the PhysicalProperties module. It ensures
that vertex data, which is fundamental to defining the shape and
characteristics of physical objects, resides in the appropriate
module.