Decouples Platform from GLFWWindow implementation
Moves towards a more abstract and flexible architecture by decoupling the Platform class from the concrete GLFWWindow implementation.

This change achieves this by introducing interfaces IWindow and IGraphicsContext and utilizing WindowGLFW and GraphicsContextGLFW as concrete implementations. This allows for easier swapping of windowing or graphics context backends in the future.