Initializes object IDs automatically
Ensures objects are assigned unique IDs upon creation.

This change introduces an atomic counter to automatically
generate and assign unique IDs to newly created objects.

Also, this fixes an issue where the object ID was not being
correctly passed to the GPU mesh creation, so it uses the
object ID instead of some other arbitrary id.